Sort layout.go Reader types and move parseCoded to coded.go
